InitialCommit
Above project contains 2 folders meant to be integrated.The webView folder contains framework 7.0 and the other folder contains 4.something. Both projects works independently correctly.However When trying to integrate there seems to be references that are not being picked up in the program even after explicitly referencing them in the file. I'm basically trying to integrate the webview application into the other project.I figured I try and create a new project(not in zip folder) and use a later framework (7.0 standard) for the wfp application.

I have tried unloading the project.Specifying the Target Framework in Dashboard assembly information.It prompts me to download the specified framework(which I did manually but ) or continue to use the successful target framework.

Webview is a simple app that displays a website by specifying a url in the webview app.

Dashboard is the project with various buttons that is meant to display various views depending on button clicked.

Each button contains a different view that is created in the views folder.

I have a base class that all other classes of the program inherit from.